target_system/pyproject.toml
[tool.poetry]
name = "aha-target-system"
version = "0.1.0"
description = "Multi-Agent Research Team - Target System for AHA Demo"
authors = ["Gemini <gemini@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
zephyr-agents = "^0.1.0"
langsmith = "^0.0.69"
tavily-python = "^0.3.0"
openai = "^1.3.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
